Maybe we don't need unit tests here?
That sounds so incredibly foolish, but we've got a pretty thin client here with a deep dependency on an external API and is mostly front-end facing. Kind of a worst case scenario for unit testing, right? Maximum cost and minimum benefit. Unit tests on the design system and the API, for sure, but this part doesn't do much more than pass things along to one or the other. Still, I figure there will come a day when I regret this, but I'm also pretty sure it will be better to address that when that day comes and we know exactly the shape of the problem we're facing, than to spin my wheels a bunch right now trying to anticipate what those future gaps might be. Sorry, Future Jason, if I try to solve all of your problems you'll still be trying to get this damn thing out in the first place.